requirements.txt
pypdf
pdf2image
#typer[all]
torch
torchvision
#transformers
qwen-vl-utils
#accelerate
autoawq
#poppler required for pdf2image